pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iottrafficgen"
version = "0.1.0"
description = "Reproducible IoT traffic generation framework for controlled laboratories"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "GICAP Research Group"}]
keywords = ["iot", "cybersecurity", "traffic-generation", "reproducibility"]

dependencies = [
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
iottrafficgen = "iottrafficgen.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]